// FoodDiary.Web.Client/src/app/shared/api/sdk/generated/model/shopping-list-http-response.ts
/**
 * FoodDiary User API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ShoppingListItemHttpResponse } from './shopping-list-item-http-response';

export interface ShoppingListHttpResponse {
    createdAt?: string;
    id?: string;
    items?: Array<ShoppingListItemHttpResponse>;
    name?: string;
}