// FoodDiary.Web.Client/src/app/shared/api/sdk/generated/model/shopping-list-overview-http-response.ts
/**
 * FoodDiary User API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ShoppingListHttpResponse } from './shopping-list-http-response';
import { ShoppingListPageHttpResponse } from './shopping-list-page-http-response';

export interface ShoppingListOverviewHttpResponse {
    lists?: ShoppingListPageHttpResponse;
    selectedList?: ShoppingListHttpResponse | null;
}